+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+2.4 [25th of September 2020]
+- Allow for storing common environments' configuration in /configuration/common.php which is executed AFTER the environment specific config
+- use Dotenv::createUnsafeImmutable() for legacy reasons, we should stick to $_ENV though
+
 2.3.1 [24th of September 2020]
 - change usage of getenv to $_ENV. Potentially dangerous, needs investigation.